renaming spacing tokens
