+++ ff.web/src/components/pages/ProfilePage.tsx
@@ -16,7 +16,7 @@ const ProfilePage = function (props: any) {
             block
             variant="large"
             type="submit" 
-            color="secondary"
+            color="primary"
             onClick={() => {
               toggleTheme()
             }}
