+++ ff.web/src/components/templates/Page.module.scss
@@ -1,0 +1,10 @@
+.container {
+    // border: 1px solid cyan;
+    display: flex;
+    flex: 1;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: var(--space-16);
+  }
+
