+++ ff.web/src/scss/_base.scss
@@ -17,7 +17,7 @@ body {
 
   margin: 0;
   color: var(--text-primary);
-  background: var(--surface-primary);
+  background: var(--surface-secondary);
 }
 
 a {
