+++ ff.web/src/scss/_layout.scss
@@ -1,27 +1,27 @@
 :root {
   // Spacing variables
-  --space-xxs: #{$space-xxs};
-  --space-xs: #{$space-xs};
-  --space-sm: #{$space-sm};
-  --space-md: #{$space-md};
-  --space-lg: #{$space-lg};
-  --space-xl: #{$space-xl};
-  --space-xxl: #{$space-xxl};
-  --space-3xl: #{$space-3xl};
-  --space-4xl: #{$space-4xl};
-  --space-5xl: #{$space-5xl};
-  --space-6xl: #{$space-6xl};
-  --space-7xl: #{$space-7xl};
-  --space-8xl: #{$space-8xl};
+  --space-2: #{$space-2}; // xxs
+  --space-4: #{$space-4}; // xs
+  --space-8: #{$space-8}; // sm
+  --space-16: #{$space-16}; // md
+  --space-24: #{$space-24}; // lg
+  --space-32: #{$space-32}; // xl
+  --space-40: #{$space-40}; // xxl
+  --space-48: #{$space-48}; // 3xl
+  --space-64: #{$space-64}; // 4xl
+  --space-80: #{$space-80}; // 5xl
+  --space-88: #{$space-88}; // 6xl
+  --space-96: #{$space-96}; // 7xl
+  --space-104: #{$space-104}; // 8xl
 
   // Radius variables
-  --radius-xs: #{$radius-xs};
-  --radius-sm: #{$radius-sm};
-  --radius-md: #{$radius-md};
-  --radius-lg: #{$radius-lg};
-  --radius-xl: #{$radius-xl};
-  --radius-xxl: #{$radius-xxl};
-  --radius-3xl: #{$radius-3xl};
+  --radius-2: #{$radius-2}; // xs
+  --radius-4: #{$radius-4}; // sm
+  --radius-8: #{$radius-8}; // md
+  --radius-16: #{$radius-16}; // lg
+  --radius-24: #{$radius-24}; // xl
+  --radius-32: #{$radius-32}; // xxl
+  --radius-128: #{$radius-128}; // 3xl
 }
 
 html {
@@ -31,12 +31,6 @@ html {
 }
 
 #root {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-}
-
-.app {
   flex: 1;
   display: flex;
   flex-direction: column;
