+++ ff.web/src/scss/_tokens.scss
@@ -36,25 +36,25 @@ $button-text: #FFFFFF;
 
 //SPACING TOKENS
 
-$space-xxs: 2px; 
-$space-xs: 4px;
-$space-sm: 8px; 
-$space-md: 16px; 
-$space-lg: 24px; 
-$space-xl: 32px; 
-$space-xxl: 40px;
-$space-3xl: 48px; 
-$space-4xl: 64px; 
-$space-5xl: 80px;
-$space-6xl: 88px;
-$space-7xl: 96px;
-$space-8xl: 104px;
+$space-2: 2px; // xxs
+$space-4: 4px; // xs
+$space-8: 8px; // sm
+$space-16: 16px; // md
+$space-24: 24px; // lg
+$space-32: 32px; // xl
+$space-40: 40px; // xxl
+$space-48: 48px; // 3xl
+$space-64: 64px; // 4xl
+$space-80: 80px; // 5xl
+$space-88: 88px; // 6xl
+$space-96: 96px; // 7xl
+$space-104: 104px; // 8xl
 
-//RADIUS TOKENS
-$radius-xs: 2px;
-$radius-sm: 4px;
-$radius-md: 8px;
-$radius-lg: 16px;
-$radius-xl: 24px;
-$radius-xxl: 32px;
-$radius-3xl: 128px;
+// RADIUS TOKENS
+$radius-2: 2px; // xs
+$radius-4: 4px; // sm
+$radius-8: 8px; // md
+$radius-16: 16px; // lg
+$radius-24: 24px; // xl
+$radius-32: 32px; // xxl
+$radius-128: 128px; // 3xl
